pyproject.toml
[tool.poetry]
name = "foxops"
version = "0.1.0"
description = "GitOps for GitLab. Maintain your repositories and configuration as-code."
authors = ["Ariel Richtman <10679234+arichtman@users.noreply.github.com>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.9"
python-gitlab = "^2.10.1"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
mkdocs = "^1.2.2"
black = "^21.9b0"
commitizen = "^2.18.1"
pre-commit = "^2.15.0"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
